Remove manual process management by using CommandContext()
Golang provides the function exec.CommandContext(). Using this, a
`cancel()` function can be defered and dangling ressources are cleaned
up automatically at function exit.

Using this the code the cleanup logic can be unified and everything
stays at the same location in the code. The concept is described in this
blogpost: https://blog.golang.org/context